decls is the new index
